requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pandas==2.2.2
pyyaml==6.0.1
numpy==1.26.4
regex==2024.5.15
python-dateutil==2.9.0.post0
pytest==8.0.0
langchain==0.2.14
langchain-community==0.2.12
openai==1.54.3
langchain-openai==0.1.0
tiktoken==0.7.0
python-dotenv==1.0.1